requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.39.0
datasets>=2.18.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.5.1
speechbrain>=1.0.0
jiwer>=3.0.3
pyyaml>=6.0.1
streamlit>=1.33.0
matplotlib>=3.7.0
pytest>=8.0.0
